修改calculate_gcd函数签名：将参数类型从int改为long

+++ math_utils.c
@@ -3,7 +3,7 @@
 #include "math_utils.h"
 
 // 目标追溯函数：计算两个数的最大公约数
-int calculate_gcd(int a, int b) {
+long calculate_gcd(long a, long b) {
     // 添加输入验证
     if (a < 0) a = -a;
     if (b < 0) b = -b;
@@ -16,7 +16,7 @@ int calculate_gcd(int a, int b) {
 
 // 计算两个数的最小公倍数
 int calculate_lcm(int a, int b) {
-    return (a * b) / calculate_gcd(a, b);
+    return (a * b) / (int)calculate_gcd(a, b);
 }
 
 // 判断是否为质数
